+++ src/ollama_mcp/__init__.py
@@ -9,19 +9,17 @@ Author: Paolo Dalprato
 License: MIT
 """
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"
 __author__ = "Paolo Dalprato"
 __email__ = "paolo@paolodalprato.com"
 __license__ = "MIT"
 
-from .client import OllamaClient  # Use fixed client.py version
-from .server_manager import OllamaServerManager
+from .client import OllamaClient
 from .model_manager import ModelManager
 from .hardware_checker import HardwareChecker
 
 __all__ = [
     "OllamaClient",
-    "OllamaServerManager", 
     "ModelManager",
     "HardwareChecker"
 ]